Rename RDL package name
