+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-binder/src/main/java/org/apache/shardingsphere/distsql/parser/binder/context/CreateShardingRuleStatementContext.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.distsql.parser.binder.context;
+
+import org.apache.shardingsphere.distsql.parser.binder.util.ShardingAlgorithmPropertiesUtil;
+import org.apache.shardingsphere.distsql.parser.statement.rdl.CreateShardingRuleStatement;
+import org.apache.shardingsphere.distsql.parser.statement.rdl.TableRuleSegment;
+import org.apache.shardingsphere.infra.binder.statement.CommonSQLStatementContext;
+
+import java.util.Properties;
+
+/**
+ * Create sharding rule statement context.
+ */
+public final class CreateShardingRuleStatementContext extends CommonSQLStatementContext<CreateShardingRuleStatement> {
+    
+    public CreateShardingRuleStatementContext(final CreateShardingRuleStatement sqlStatement) {
+        super(sqlStatement);
+    }
+    
+    /**
+     * Get algorithm properties.
+     *
+     * @param segment segment
+     * @return algorithm properties
+     */
+    public Properties getAlgorithmProperties(final TableRuleSegment segment) {
+        return ShardingAlgorithmPropertiesUtil.getProperties(segment.getAlgorithmType(), segment.getProperties());
+    }
+}
