+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/java/org/apache/shardingsphere/distsql/parser/engine/ShardingSphereSQLStatementParserEngine.java
@@ -1,0 +1,52 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.distsql.parser.engine;
+
+import org.apache.shardingsphere.distsql.parser.engine.engine.RDLSQLStatementParserEngine;
+import org.apache.shardingsphere.infra.parser.SQLStatementParserEngine;
+import org.apache.shardingsphere.infra.parser.standard.StandardSQLStatementParserEngine;
+import org.apache.shardingsphere.infra.parser.standard.StandardSQLStatementParserEngineFactory;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
+
+/**
+ * ShardingSphere SQL statement parser engine.
+ */
+public final class ShardingSphereSQLStatementParserEngine implements SQLStatementParserEngine {
+    
+    private final StandardSQLStatementParserEngine standardSQLStatementParserEngine;
+    
+    private final RDLSQLStatementParserEngine rdlSQLStatementParserEngine;
+    
+    public ShardingSphereSQLStatementParserEngine(final String databaseTypeName) {
+        standardSQLStatementParserEngine = StandardSQLStatementParserEngineFactory.getSQLStatementParserEngine(databaseTypeName);
+        rdlSQLStatementParserEngine = new RDLSQLStatementParserEngine();
+    }
+    
+    @Override
+    public SQLStatement parse(final String sql, final boolean useCache) {
+        SQLStatement result;
+        try {
+            result = standardSQLStatementParserEngine.parse(sql, useCache);
+            // CHECKSTYLE:OFF
+        } catch (final Exception ex) {
+            // CHECKSTYLE:ON
+            result = rdlSQLStatementParserEngine.parse(sql, useCache);
+        }
+        return result;
+    }
+}
